% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckd.data}
\alias{ckd.data}
\title{A synthetic dataset contains variables for eGFR calculation.}
\format{
A data frame with 2000 rows (1000 adults and 1000 children/young adults) and 12 variables:
\describe{
  \item{cr}{Serum creatinine, micromol/L}
  \item{cys}{Serum cystatin C, mg/L}
  \item{age}{Age, years}
  \item{sex}{Sex}
  \item{ethnicity}{Ethnicity}
  \item{height}{Height, cm}
  \item{category}{Indication on whether the generated data refer to adults or children}
}
}
\source{
Synthetic dataset was generated based on two publications:  
\itemize{
  \item adults: Lamb EJ, Barratt J, Brettell EA et al. Accuracy of glomerular filtration rate estimation using creatinine and cystatin C for identifying and monitoring moderate chronic kidney disease: the eGFR-C study.  Health Technol Assess 2024;28(35), doi:10.3310/HYHN1078.  
  \item children/young adults: Pierce CB, Muñoz A, Ng DK et al. Age- and sex-dependent clinical equations to estimate glomerular filtration rates in children and young adults with chronic kidney disease. Kidney International. 2021;99(4):948–956, doi:10.1016/j.kint.2020.10.047.
}
}
\usage{
ckd.data
}
\description{
A synthetic dataset contains variables for eGFR calculation for 1000 adults and 1000 children.
}
\details{
Citation: Bikbov B. kidney.epi: Kidney-Related Functions for Clinical and Epidemiological Research. Scientific-Tools.Org, https://Scientific-Tools.Org. DOI: 10.32614/CRAN.package.kidney.epi
}
\keyword{datasets}
