% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr_ckdepi2021.R
\name{egfr.ckdepi.cr}
\alias{egfr.ckdepi.cr}
\title{Alias to the latest eGFR CKD-EPI creatinine-based equation}
\usage{
egfr.ckdepi.cr(...)
}
\arguments{
\item{...}{all arguments for the egfr.ckdepi.cr.2021 function.}
}
\value{
numeric eGFR expressed in ml/min/1.73m\ifelse{html}{\out{<sup>2</sup>}}{\eqn{^2}}.
}
\description{
Alias to the latest eGFR CKD-EPI creatinine-based equation
}
\details{
The function is just an alias to the latest eGFR CKD-EPI creatinine-based equation.

Citation: Bikbov B. kidney.epi: Kidney-Related Functions for Clinical and Epidemiological Research. Scientific-Tools.Org, https://Scientific-Tools.Org. DOI: 10.32614/CRAN.package.kidney.epi
}
