% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidney.epi.R
\name{kidney.epi}
\alias{kidney.epi}
\title{Kidney functions for R: clinical and epidemiological}
\description{
Package contains different functions for use in the field of kidney disease and general epidemiology
Current version contains functions for calculation of estimated GFR by different equations, and calculation of KDPI and KDRI for kidney transplant donors.
More functions are coming soon.
}
