% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktx.kdpi.optn.R
\name{ktx.kdpi.optn.show.years}
\alias{ktx.kdpi.optn.show.years}
\title{Shows which years are available in the R package for the OPTN mapping table, KDRI scaling factor, etc.}
\usage{
ktx.kdpi.optn.show.years()
}
\value{
numeric List of years which could be used for the argument mapping_values_year of the ktx.kdpi.optn function.
}
\description{
Shows which years are available in the R package for the OPTN mapping table, KDRI scaling factor, etc.
}
\details{
Service function which shows for user for which year(s) the OPTN mapping table, as well as KDRI scaling factor and chances of hypertension and diabetes in case if they were unknown for donor in the ktx.kdpi_mapping_table and ktx.kdpi_coefficients_table. This years could be used for the argument \emph{mapping_values_year} of the ktx.kdpi.optn function.

This function has no arguments.
}
