% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.get_named_matrix_value}
\alias{matrix.get_named_matrix_value}
\title{Access matrix cells by row and column names}
\usage{
matrix.get_named_matrix_value(matrix_data, row_name, col_name)
}
\arguments{
\item{matrix_data}{matrix. Matrix with data.}

\item{row_name}{character. Row name in the matrix.}

\item{col_name}{character. Column name in the matrix.}
}
\value{
vector. Matrix value.
}
\description{
Access matrix cells by row and column names
}
\details{
Access matrix cells by row and column names.
}
