% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.read_excel_to_named_matrix}
\alias{matrix.read_excel_to_named_matrix}
\title{Read Excel file and convert it to matrix with row and column names.}
\usage{
matrix.read_excel_to_named_matrix(file_path, sheet_name = NULL)
}
\arguments{
\item{file_path}{character. Path to Excel file.}

\item{sheet_name}{character. Name of Excel sheet. Optional, if there is only one sheet with data, the function will read it with no need to specifying the sheet name.}
}
\value{
Excel file saved to a specified folder.
}
\description{
Read Excel file and convert it to matrix with row and column names.
}
\details{
Read Excel file and convert it to matrix with row and column names.
}
