% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd_classify.R
\name{nephro.albuminuria_category}
\alias{nephro.albuminuria_category}
\title{Legacy function to calculate albuminuria categories}
\usage{
nephro.albuminuria_category(...)
}
\arguments{
\item{...}{all arguments for the ckd.kdigo_category.albuminuria() function.}
}
\description{
Legacy function to calculate albuminuria categories
}
\details{
Legacy function to calculate albuminuria categories (A1, A2, A3) based on albuminuria values. Alias to the ckd.kdigo_category.albuminuria() function. This function is for legacy use only.
}
