% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd_classify.R
\name{nephro.gfr_category}
\alias{nephro.gfr_category}
\alias{nephro.kdigo_risk_category}
\title{Legacy function to calculate KDIGO GFR categories}
\usage{
nephro.gfr_category(...)

nephro.kdigo_risk_category(...)
}
\arguments{
\item{...}{all arguments for the ckd.kdigo_category.risk() function.}
}
\description{
Legacy function to calculate KDIGO GFR categories

Legacy function to calculate KDIGO risk categories
}
\details{
Legacy function to calculate KDIGO GFR categories. Alias to the ckd.kdigo_category.gfr() function. This function is for legacy use only.

Legacy function to calculate KDIGO risk categories. Alias to the ckd.kdigo_category.risk() function. This function is for legacy use only.
}
