% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd_classify.R
\name{nephro.proteinuria_category}
\alias{nephro.proteinuria_category}
\title{Legacy function to calculate proteinuria categories}
\usage{
nephro.proteinuria_category(...)
}
\arguments{
\item{...}{all arguments for the ckd.kdigo_category.proteinuria() function.}
}
\description{
Legacy function to calculate proteinuria categories
}
\details{
Legacy function to calculate albuminuria categories (A1, A2, A3) based on proteinuria values. Alias to the ckd.kdigo_category.proteinuria() function. This function is for legacy use only.
}
