% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.general.R
\name{service.check_param_arguments}
\alias{service.check_param_arguments}
\title{Check the argument of a given parameter which set by user and stop function if the value set by user is not among the possible values of the argument}
\usage{
service.check_param_arguments(param2check, possible_params,
  custom_message = "")
}
\arguments{
\item{param2check}{List, Character string, Number. Parameter used in a function.}

\item{possible_params}{List. List of possible values of the parameter arguments}

\item{custom_message}{Character string. Custom message to be output. If not defined, the standart output message is provided.
nothing to return}
}
\description{
Check the argument of a given parameter which set by user and stop function if the value set by user is not among the possible values of the argument
}
\details{
Check the argument of a given parameter whichset by user and stop function if the value set by user is not among the possible values of the argument.
Service function that will not be exported to user.

Programming: Boris Bikbov \email{boris@bikbov.ru}.
}
