% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.general.R
\name{service.check_param_number}
\alias{service.check_param_number}
\title{Check number of parameters and stop function if it exceeds the expected number of parameters}
\usage{
service.check_param_number(param2check, acceptable_number = 1,
  custom_message = "")
}
\arguments{
\item{param2check}{List, Character string, Number. Parameter used in a function.}

\item{acceptable_number}{Numeric. Acceptable number of arguments in the list param2check (by default is "1")}

\item{custom_message}{Character string. Custom message to be output. If not defined, the standart output message is provided.
nothing to return}
}
\description{
Check number of parameters and stop function if it exceeds the expected number of parameters
}
\details{
Check number of parameters and stop function if it exceeds the expected number of parameters.
Service function that will not be exported to user.

Programming: Boris Bikbov \email{boris@bikbov.ru}.
}
