% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.general.R
\name{service.check_params_numeric}
\alias{service.check_params_numeric}
\title{Check whether the following variables are numeric and stop function if at least one of them is not numeric}
\usage{
service.check_params_numeric(...)
}
\arguments{
\item{...}{Argument list. Argument list (arbitruary number of valiables) with data to check.
nothing to return}
}
\description{
Check whether the following variables are numeric and stop function if at least one of them is not numeric
}
\details{
Check whether the following variables are numeric and stop function if at least one of them is not numeric.
Service function that will not be exported to user.

Programming: Boris Bikbov \email{boris@bikbov.ru}.
}
