% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.check_plausibility.R
\name{service.check_plausibility.age}
\alias{service.check_plausibility.age}
\title{Service functions for data check on biological plausibility and biochemistry conversion which could be applied in any function of the package or externally}
\usage{
service.check_plausibility.age(age)
}
\arguments{
\item{age}{Numeric. The value to be checked.}
}
\value{
numeric Vector with controlled values.
}
\description{
Check and modify if necessary the age values.
}
\details{
Service function which check whether age is in biologically plausible boundaries, shows to user warnings if any, and substitute unplausable values.
}
