% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.check_plausibility.R
\name{service.convert_creatinine}
\alias{service.convert_creatinine}
\title{Convert creatinine values if necessary (depending on the mesurement units).}
\usage{
service.convert_creatinine(creatinine, creatinine_units)
}
\arguments{
\item{creatinine}{Numeric. The creatinine value from data set.}

\item{creatinine_units}{Character. Creatinine mesurement units defined by user.}
}
\value{
numeric Vector with converted values.
}
\description{
Convert creatinine values if necessary (depending on the mesurement units).
}
\details{
Service function which check mesurement units and convert creatinine values if necessary.
}
