% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.check_plausibility.R
\name{service.convert_cystatin}
\alias{service.convert_cystatin}
\title{Convert cystatin C values between measurement units.}
\usage{
service.convert_cystatin(cystatin, cystatin_units)
}
\arguments{
\item{cystatin}{Numeric. The cystatin C values from a data set.}

\item{cystatin_units}{Character. Cystatin C measurement units in a data set.}
}
\value{
numeric Cystatin C values converted in mg/l.
}
\description{
Convert cystatin C values between measurement units.
}
\details{
Service function which check measurement units and convert cystatin C values to mg/l.
}
