% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.general.R
\name{service.count_lowerequal_threshhold}
\alias{service.count_lowerequal_threshhold}
\title{Count how many values are less or equal than the defined threshhold.}
\usage{
service.count_lowerequal_threshhold(x, threshhold)
}
\arguments{
\item{x}{the vector to be checked.}

\item{threshhold}{numeric the threshhold to compare with.}
}
\value{
numeric returns number of numeric values less or equal to the threshhold.
}
\description{
Count how many values are less or equal than the defined threshhold.
}
\details{
Count how many values are less or equal than the defined threshhold. 

Programming: Boris Bikbov \email{boris@bikbov.ru}.
}
\examples{
myvals <- c(1, 8, -5, "oggi", NA)
myvals2 <- service.count_lowerequal_threshhold(myvals, 0)
myvals2 # 1
}
