% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.general.R
\name{service.is_numeric}
\alias{service.is_numeric}
\title{Check whether a vector is numeric.}
\usage{
service.is_numeric(x)
}
\arguments{
\item{x}{the vector to be checked.}
}
\value{
logic whether vector x is numeric or not.
}
\description{
Check whether a vector is numeric.
}
\details{
Check whether a vector is numeric. 

Programming: Boris Bikbov \email{boris@bikbov.ru}.
}
