% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.general.R
\name{service.output_message}
\alias{service.output_message}
\title{Produce message for warning or cat}
\usage{
service.output_message(x, custom_phrase, warning_type)
}
\arguments{
\item{x}{Numeric. The value to be checked (usualy a counter of some variable).}

\item{custom_phrase}{Character string. Custom message to be inserted in the middle of standard message.}

\item{warning_type}{Character string. The type of message: warning (with substitution to NA) or cat (with leave as is).}
}
\value{
Character string. Returns a phrase.
}
\description{
Produce message for warning or cat
}
\details{
Produce message that is used by warning or cat in the ktx.kdpi.optn function. 
Service function that will not be exported to user, and used only in the ktx.kdpi.optn function.

Programming: Boris Bikbov \email{boris@bikbov.ru}.
}
