% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.general.R
\name{service.singular_or_plural}
\alias{service.singular_or_plural}
\title{Form output message in singular or plural.}
\usage{
service.singular_or_plural(x, singular, plural)
}
\arguments{
\item{x}{Numeric. The value to be checked (usualy a counter of some variable).}

\item{singular}{Character string. The value to be returned in case of singular form (usualy a string, but could be any type).}

\item{plural}{Character string. The value to be returned in case of plural form (usualy a string, but could be any type).}
}
\value{
Character string. Returns a value for constructing messages in singular or plural form.
}
\description{
Form output message in singular or plural.
}
\details{
Provide different output for constructing messages in singular or plural. 

Programming: Boris Bikbov \email{boris@bikbov.ru}.
}
\examples{
service.singular_or_plural(1, "This value was", "These values were") # "This value was"
service.singular_or_plural(99, "This value was", "These values were") # "These values were"
}
