% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr_schwartz.R
\name{egfr.schwartz}
\alias{egfr.schwartz}
\title{Deprecated function name. Use specific functions egfr.schwartz.cr(), egfr.schwartz.cys(), egfr.schwartz.cr_cys().}
\usage{
egfr.schwartz(...)
}
\arguments{
\item{...}{used for legacy only.}
}
\description{
Deprecated function name. Use specific functions egfr.schwartz.cr(), egfr.schwartz.cys(), egfr.schwartz.cr_cys().
}
\details{
Deprecated function name.
}
