% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.check_plausibility.R
\name{service.convert_creatinine}
\alias{service.convert_creatinine}
\title{Convert creatinine values between measurement units.}
\usage{
service.convert_creatinine(
  creatinine,
  creatinine_units,
  creatinine_reference_units = "mg/dl"
)
}
\arguments{
\item{creatinine}{Numeric. The creatinine value from a data set.}

\item{creatinine_units}{Character. Creatinine mesurement units in a data set.}

\item{creatinine_reference_units}{Character. Creatinine measurement units as a desired output (mg/dl by default).}
}
\value{
numeric Creatinine values converted into reference measurement units.
}
\description{
Convert creatinine values between measurement units.
}
\details{
Service function which check measurement units and convert creatinine values to selected by user.
}
